refactor : Util패키지 생성에따른 패키지변경
우승자 판단 유틸을 util패키지를 생성하여 패키지 구조변경을 하였습니다.
