+++ src/main/java/step4/util/JudgeRacingResult.java
@@ -1,4 +1,4 @@
-package step4;
+package step4.util;
 
 import step4.dto.Car;
